chore(map): rename add-instructions to add-information
