+++ map/src/components/add-information.tsx
